♻️ GoogleMaps表示用のページをリファクタリングする
