+++ frontend/src/pages/maps.tsx
@@ -1,31 +1,15 @@
-import { GoogleMap, LoadScript, Marker } from '@react-google-maps/api'
+import Link from 'next/link'
+import Map from '@/components/Map'
 
-const containerStyle = {
-  width: '400px',
-  height: '400px',
-}
-
-const center = {
-  lat: 35.693377849778024, // 緯度
-  lng: 139.7494393926594, // 経度
-}
-
-const postitionTokyoTower = {
-  lat: 35.658581,
-  lng: 139.745433,
-}
-
-const mapsApiKey = process.env.NEXT_PUBLIC_GOOGLE_MAPS_API_KEY
-
-const MyComponent = () => {
+function App() {
   return (
-    <LoadScript googleMapsApiKey={mapsApiKey || ''}>
-      <GoogleMap mapContainerStyle={containerStyle} center={center} zoom={17}>
-        <Marker position={center} />
-        <Marker position={postitionTokyoTower} />
-      </GoogleMap>
-    </LoadScript>
+    <div>
+      <div>
+        <Link href='/'>トップページへ</Link>
+      </div>
+      <Map />
+    </div>
   )
 }
 
-export default MyComponent
+export default App
